Remove empty cq4 log files to prevent BinLog from failing to start
This patch also backports CASSANDRA-17595.

patch by Stefan Miklosovic; reviewed by Caleb Rackliffe for CASSANDRA-17933

+++ src/java/org/apache/cassandra/utils/binlog/BinLog.java
@@ -20,7 +20,6 @@ package org.apache.cassandra.utils.binlog;
 
 import java.io.File;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
@@ -28,6 +27,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.function.Function;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
@@ -36,19 +36,21 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import net.openhft.chronicle.queue.ChronicleQueue;
+import net.openhft.chronicle.queue.impl.single.SingleChronicleQueue;
 import net.openhft.chronicle.queue.impl.single.SingleChronicleQueueBuilder;
 import net.openhft.chronicle.queue.ExcerptAppender;
 import net.openhft.chronicle.queue.RollCycles;
 import net.openhft.chronicle.wire.WireOut;
 import net.openhft.chronicle.wire.WriteMarshallable;
 import org.apache.cassandra.concurrent.NamedThreadFactory;
-import org.apache.cassandra.fql.FullQueryLoggerOptions;
 import org.apache.cassandra.io.FSError;
 import org.apache.cassandra.io.util.FileUtils;
 import org.apache.cassandra.utils.JVMStabilityInspector;
 import org.apache.cassandra.utils.NoSpamLogger;
 import org.apache.cassandra.utils.Throwables;
 import org.apache.cassandra.utils.concurrent.WeightedQueue;
+
+import static java.lang.String.format;
 
 /**
  * Bin log is a is quick and dirty binary log that is kind of a NIH version of binary logging with a traditional logging
@@ -426,6 +428,12 @@ public class BinLog implements Runnable
             }
             try
             {
+                Throwable sanitationThrowable = cleanEmptyLogFiles(path.toFile(), null);
+                if (sanitationThrowable != null)
+                    throw new RuntimeException(format("Unable to clean up %s directory from empty %s files.",
+                                                      path.toAbsolutePath(), SingleChronicleQueue.SUFFIX),
+                                               sanitationThrowable);
+
                 // create the archiver before cleaning directories - ExternalArchiver will try to archive any existing file.
                 BinLogArchiver archiver = Strings.isNullOrEmpty(archiveCommand) ? new DeletingArchiver(maxLogSize) : new ExternalArchiver(archiveCommand, path, maxArchiveRetries);
                 if (cleanDirectory)
@@ -462,24 +470,46 @@ public class BinLog implements Runnable
         }
     }
 
+    /**
+     * ChronicleQueue fails to start on cq4 files which are empty. Find such files in log dir and remove them.
+     */
+    private static Throwable cleanEmptyLogFiles(File directory, Throwable accumulate)
+    {
+        return cleanDirectory(directory, accumulate,
+                              (dir) -> dir.listFiles(file -> {
+                                  boolean foundEmptyCq4File = !file.isDirectory()
+                                                              && file.length() == 0
+                                                              && file.getName().endsWith(SingleChronicleQueue.SUFFIX);
+
+                                  if (foundEmptyCq4File)
+                                      logger.warn("Found empty ChronicleQueue file {}. This file wil be deleted as part of BinLog initialization.",
+                                                  file.getAbsolutePath());
+
+                                  return foundEmptyCq4File;
+                              }));
+    }
+
     public static Throwable cleanDirectory(File directory, Throwable accumulate)
     {
-        if (!directory.exists())
-        {
-            return Throwables.merge(accumulate, new RuntimeException(String.format("%s does not exists", directory)));
-        }
-        if (!directory.isDirectory())
-        {
-            return Throwables.merge(accumulate, new RuntimeException(String.format("%s is not a directory", directory)));
-        }
-        for (File f : directory.listFiles())
-        {
-            accumulate = deleteRecursively(f, accumulate);
-        }
+        return cleanDirectory(directory, accumulate, File::listFiles);
+    }
+
+    private static Throwable cleanDirectory(File directory, Throwable accumulate, Function<File, File[]> lister)
+    {
+        accumulate = checkDirectory(directory, accumulate);
+
+        if (accumulate != null)
+            return accumulate;
+
+        File[] files = lister.apply(directory);
+
+        if (files != null)
+            for (File f : files)
+                accumulate = deleteRecursively(f, accumulate);
+
         if (accumulate instanceof FSError)
-        {
             JVMStabilityInspector.inspectThrowable(accumulate);
-        }
+
         return accumulate;
     }
 
@@ -487,11 +517,22 @@ public class BinLog implements Runnable
     {
         if (fileOrDirectory.isDirectory())
         {
-            for (File f : fileOrDirectory.listFiles())
-            {
-                accumulate = FileUtils.deleteWithConfirm(f, accumulate);
-            }
+            File[] files = fileOrDirectory.listFiles();
+            if (files != null)
+                for (File f : files)
+                    accumulate = FileUtils.deleteWithConfirm(f, accumulate);
         }
         return FileUtils.deleteWithConfirm(fileOrDirectory, accumulate);
+    }
+
+    private static Throwable checkDirectory(File directory, Throwable accumulate)
+    {
+        if (!directory.exists())
+            accumulate = Throwables.merge(accumulate, new RuntimeException(format("%s does not exist", directory)));
+
+        if (!directory.isDirectory())
+            accumulate = Throwables.merge(accumulate, new RuntimeException(format("%s is not a directory", directory)));
+
+        return accumulate;
     }
 }
